docs: Add routing config for sparql endpoint and storage
Due to key/value stores using JSON objects,
the SPARQL resource store can not be used to store them,
meaning the only out of the box solution that works is memory storage.
This config makes it so file storage can be used for internal storage
while still having a SPARQL backend.