feat: Add storage modification types [CREATED, CHANGED, DELETED]
Preparaton for implementation of Solid Notifications according to https://solid.github.io/notifications/protocol
The proposal refers to https://www.w3.org/TR/activitystreams-core/ for the activity types.
In order to provide meaningful notifications to subscribers we have differentiated between the types of resource changes, such as CREATED, CHANGED and DELETED.
This required an introduction of a ModificationResult interface that wraps the changed ResourceIdentifiers.