test: Disable notifications when testing the file locker
There is a conflict between notifications being sent out right before a test stops,
and the file locker trying to clean up the locks folder.
This only has an impact when shutting down the server so has no real impact,
but should still be fixed eventually so we can easily write tests.